Better Error Handling (#12)
* Moved some files around

* Refactored the back-end
