+++ client/src/app/guards/auth.guard.ts
@@ -1,7 +1,7 @@
 import { CanActivateFn, Router, UrlTree } from '@angular/router';
 import { AuthService } from '../services/auth.service';
 import { inject } from '@angular/core';
-import { Observable, map, of } from 'rxjs';
+import { Observable, catchError, map, of } from 'rxjs';
 
 export const authGuard: CanActivateFn = (route, state): Observable<boolean | UrlTree> => {
     const path: string = route.url[0]?.path;
@@ -9,24 +9,19 @@ export const authGuard: CanActivateFn = (route, state): Observable<boolean | Url
     const router: Router = inject(Router);
 
     return auth.hasAuthToken().pipe(
-        map((resp: any) => {
-            let allow = true;
-            let url = '/login';
-
-            if(!resp.body.hasToken) {
-                allow = false;
-            }
-
+        map(() => {
             if(path === undefined || path === 'login') {
-                allow = !allow;
-                url = '/home';
+                return router.parseUrl('/home');
             }
 
-            if(allow) {
-                return true;
+            return true;
+        }),
+        catchError(() => {
+            if(path === undefined || path === 'login') {
+                return of(true);
             }
 
-            return router.parseUrl(url);
-        })
+            return of(router.parseUrl('/login'));
+        }),
     );
 };
