+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "express-async-handler": "^1.2.0"
+  }
+}
