+++ server/routes/auth.js
@@ -1,30 +1,10 @@
 const router = require('express').Router();
-const { SpotifyAuth } = require('../spotify/spotifyAuth.js');
+const { SpotifyAuth } = require('../controllers/spotify/spotifyAuth.js');
 
-router.get('/login', async (req, res) => {
-    res.send({ url: await SpotifyAuth.generateSpotifyAuthLink() });
-});
-
-router.get('/refresh', async (req, res) => {
-    await SpotifyAuth.refreshCurrentTokens(req, res);
-
-    res.send('Refreshed your tokens!');
-});
-
-router.get('/logout', (req, res) => {
-    SpotifyAuth.deleteCookieTokens(res);
-
-    res.send({ loggedOut: true });
-});
-
-router.get('/info', async (req, res) => {
-    await SpotifyAuth.getAuthInfo(req, res);
-
-    res.json({ authSuccessful: true });
-});
-
-router.get('/hasAuthToken', (req, res) => {
-    res.json({ hasToken: SpotifyAuth.hasAuthToken(req) });
-});
+router.get('/login', SpotifyAuth.getSpotifyAuthLink);
+router.get('/refresh', SpotifyAuth.refreshCurrentTokens);
+router.get('/logout', SpotifyAuth.deleteTokens);
+router.get('/info', SpotifyAuth.getAuthInfo);
+router.get('/hasAuthToken', SpotifyAuth.hasAuthToken);
 
 module.exports = router;
