+++ server/routes/track.js
@@ -1,10 +1,6 @@
 const router = require('express').Router();
-const { LyricsAPI } =  require('../lyrics/lyricsApi.js');
+const { LyricsAPI } =  require('../controllers/lyrics/lyricsApi.js');
 
-router.get('/lyrics', async (req, res) => {
-    const lyrics = await LyricsAPI.getLyrics(req, res);
-
-    res.json(lyrics);
-});
+router.get('/lyrics', LyricsAPI.getLyrics);
 
 module.exports = router;
