--- server/spotify/spotifyAuth.js
@@ -1,186 +1,0 @@
-const crypto = require('crypto');
-
-const clientID = process.env.SPOTIFY_CLIENT_ID;
-const clientSecret = process.env.SPOTIFY_CLIENT_SECRET;
-const clientURL = process.env.CLIENT_URL;
-const redirectURI = clientURL + '/auth';
-const domain = process.env.DOMAIN;
-
-class SpotifyAuth {
-    static hasAuthToken(req) {
-        if (req.cookies.authToken) {
-            return true;
-        }
-        return false;
-    }
-
-    static async generateSpotifyAuthLink() {
-        const state = generateRandomCode(128);
-        const hashed = await generateCodeChallenge(state);
-        const codeChallenge = base64Encode(hashed);
-
-        const scope =
-            'user-read-private user-read-email user-top-read user-read-playback-state';
-
-        const params = new URLSearchParams({
-            response_type: 'code',
-            client_id: clientID,
-            scope: scope,
-            state: state,
-            code_challenge_method: 'S256',
-            code_challenge: codeChallenge,
-            redirect_uri: redirectURI,
-        });
-
-        return 'https://accounts.spotify.com/authorize?' + params.toString();
-    }
-
-    static async getAuthInfo(req, res) {
-        const params = new URLSearchParams({
-            grant_type: 'authorization_code',
-            code: req.query.code,
-            client_id: clientID,
-            redirect_uri: redirectURI,
-            code_verifier: req.query.state,
-        });
-
-        const result = await fetch('https://accounts.spotify.com/api/token', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-            body: params,
-        });
-
-        res.status(result.status);
-
-        if(result.status === 200) {
-            const authInfo = await result.json();
-            this.setCookieTokens(authInfo, res);
-        }
-    }
-
-    static async refreshCurrentTokens(req, res) {
-        const params = new URLSearchParams({
-            grant_type: 'refresh_token',
-            refresh_token: req.cookies.refreshToken,
-            client_id: clientID,
-            client_secret: clientSecret,
-        });
-
-        const result = await fetch('https://accounts.spotify.com/api/token', {
-            method: 'POST',
-            headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-            body: params,
-        });
-    
-        res.status(result.status);
-
-        if(result.status === 200) {
-            const newAuthInfo = await result.json();
-            this.setCookieTokens(newAuthInfo, res);
-        }
-    }
-
-    static async deleteCookieTokens(res) {
-        const cookieOptions = {
-            httpOnly: true,
-            domain: domain,
-            path: '/',
-        };
-
-        res.clearCookie('authToken', cookieOptions);
-        res.clearCookie('refreshToken', cookieOptions);
-    }
-
-    static setCookieTokens(authInfo, res) {
-        const cookieOptions = {
-            httpOnly: true,
-            domain: domain,
-            path: '/',
-            maxAge: 1000 * 60 * 60 * 24 * 7,
-        };
-
-        res.cookie('authToken', authInfo.access_token, cookieOptions);
-        res.cookie('refreshToken', authInfo.refresh_token, cookieOptions);
-    }
-}
-
-function generateRandomCode(length) {
-    var text = '';
-    var possible =
-        'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-
-    for (var i = 0; i < length; i++) {
-        text += possible.charAt(Math.floor(Math.random() * possible.length));
-    }
-
-    return text;
-}
-
-async function generateCodeChallenge(plain) {
-    const encoder = new TextEncoder();
-    const data = encoder.encode(plain);
-
-    const hash = crypto.createHash('sha256');
-    hash.update(data);
-    const test = hash.digest('hex');
-    const buffer = Buffer.from(test, 'hex');
-    const challenge = buffer.buffer.slice(buffer.byteOffset, buffer.byteOffset + buffer.byteLength);
-
-    return challenge;
-}
-
-function base64Encode(input) {
-    return btoa(String.fromCharCode(...new Uint8Array(input)))
-        .replace(/=/g, '')
-        .replace(/\+/g, '-')
-        .replace(/\//g, '_');
-}
-
-// async function authenticate(req, res, next) {
-//     const refreshToken = req.cookies.refreshToken;
-//     const accessToken = req.cookies.accessToken;
-//     let tokenExpiration = req.cookies.tokenExpiration;
-
-//     const { accessToken: newAccessToken, refreshToken: newRefreshToken, tokenExpiration: newTokenExpiration } = await refreshOrGetAccessToken(refreshToken, accessToken, tokenExpiration);
-
-//     if (newAccessToken !== accessToken || newRefreshToken !== refreshToken || newTokenExpiration !== tokenExpiration) {
-//         res.cookie('accessToken', newAccessToken, { maxAge: newTokenExpiration - Date.now() / 1000 });
-//         res.cookie('refreshToken', newRefreshToken);
-//         res.cookie('tokenExpiration', newTokenExpiration);
-//     }
-
-//     req.accessToken = newAccessToken;
-//     req.refreshToken = newRefreshToken;
-//     req.tokenExpiration = newTokenExpiration;
-
-//     next();
-// }
-
-// function isTokenExpired(tokenExpiration) {
-//     return tokenExpiration < Date.now() / 1000;
-// }
-
-// async function refreshOrGetAccessToken(refreshToken, accessToken, tokenExpiration) {
-//     if (!accessToken || isTokenExpired(tokenExpiration)) {
-//         const { access_token, expires_in, refresh_token } = await SpotifyAuth.refreshCurrentTokens(refreshToken);
-//         const newTokenExpiration = Date.now() / 1000 + expires_in;
-//         return { accessToken: access_token, refreshToken: refresh_token, tokenExpiration: newTokenExpiration };
-//     } else {
-//         return { accessToken, refreshToken, tokenExpiration };
-//     }
-// }
-
-// async function getUserData(req, res) {
-//     try {
-//         const { accessToken, refreshToken, tokenExpiration } = req;
-
-//         res.status(200).json({ message: "API call successful" });
-//     } catch (error) {
-//         console.error("Error making API call:", error);
-//         res.status(500).json({ error: "Internal server error" });
-//     }
-// }
-
-module.exports = {
-    SpotifyAuth,
-};
